main/postmarketos-update-depthcharge-kernel: use kern_guid instead of label (MR 4066)
This allows to flash correct partition when having 2 installations of pmOS.

This will also avoid flashing kernel image when using U-Boot.

+++ main/postmarketos-update-depthcharge-kernel/update-depthcharge-kernel.sh
@@ -9,9 +9,13 @@ deviceinfo_cgpt_kpart=""
 # shellcheck disable=SC1091
 . /etc/deviceinfo
 
-partition=$(findfs PARTLABEL=pmOS_kernel)
-
-echo "Flashing $deviceinfo_cgpt_kpart to $partition"
-dd if="$deviceinfo_cgpt_kpart" of="$partition"
+# shellcheck disable=SC2013
+for x in $(cat /proc/cmdline); do
+	[ "$x" = "${x#kern_guid=}" ] && continue
+	GUID="${x#kern_guid=}"
+	partition=$(findfs PARTUUID="$GUID")
+	echo "Flashing $deviceinfo_cgpt_kpart to $partition"
+	dd if="$deviceinfo_cgpt_kpart" of="$partition"
+done
 
 echo "Done."
